Version 1.00 - Ready to use for testing
Changelog:
- Fixed not scoring after time out
- Changed showPic call algorythm
- Added unravel final scores on maindata.txt
- Partially fixed bug where the timeout function would be called too
soon. It can still be replicated by going too fast, but the window is
much smaller so it should be less likely to happen. However, it will not
crash the program at the end, allowing to still fully score.

Feedback was at the end not included due to the time restrains and the
need to change reasonably the code. Thus instructions must be well
explained and supervision is necessary at the practice blocks for an
accurate recording.

Explanation for the code to be included